fix(backend,web): 보고서 기타 3종 SF 레거시 정합 수정
행사사원 목표 대비 실적:
- 실적금액: 의미오류 공식 필드(dkDailyActualSalesAmount) → SF Report 총 실적
  컬럼 정합 dailyTotalActualSalesAmount(대표금액+기타금액)로 교체
- 목표금액: 자체 targetAmount → SF DailyTargetAmount formula 재현
  dkDailyTargetAmount(목표갯수×기준단가)로 교체 (DTO Long → BigDecimal)
- 거래처코드: account.branchCode → externalKey(SAP 거래처코드)
- 전문행사조: 사원 마스터 현재 값 → 조원일정 투입 당시 값

전문행사조 확정 인원:
- 지점명: 거래처 소재 지점 → 사원 소속 지점(SF BranchName__c = FullName__r.OrgName)

진열사원 유효사원(확정):
- 목록 size 상한 100 → 10,000 (SF Report 전건 표시 정합 — 100건 절삭으로
  화면·엑셀이 잘리던 결함 해소), 웹 요청 size 10000

+++ backend/src/main/kotlin/com/otoki/powersales/domain/activity/promotion/dto/response/PromotionTargetActualReportResponses.kt
@@ -12,7 +12,7 @@ data class PromotionTargetActualReportResponse(
     val startDate: String,
     val endDate: String,
     val groups: List<PromotionTargetActualReportGroup>,
-    val totalTargetAmount: Long,
+    val totalTargetAmount: BigDecimal,
     val totalActualAmount: BigDecimal,
     val totalPrimaryQuantity: BigDecimal,
     val totalOtherQuantity: BigDecimal,
@@ -22,7 +22,7 @@ data class PromotionTargetActualReportResponse(
 /** 행사명 그룹 — rows + 그룹 소계. */
 data class PromotionTargetActualReportGroup(
     val promotionName: String?,
-    val subtotalTargetAmount: Long,
+    val subtotalTargetAmount: BigDecimal,
     val subtotalActualAmount: BigDecimal,
     val subtotalPrimaryQuantity: BigDecimal,
     val subtotalOtherQuantity: BigDecimal,
@@ -38,8 +38,9 @@ data class PromotionTargetActualChartItem(
 /**
  * 목표/실적 1행 (23컬럼) — promotionEmployee × promotion × account × product × employee × teamMemberSchedule.
  *
- * actualAmount = SF Formula `DailyActualSalesAmount__c` 재현(dkDailyActualSalesAmount 파생).
- * isWorkReport/commuteDate 는 teamMemberSchedule 조인.
+ * targetAmount = SF Formula `DKRetail__DailyTargetAmount__c` 재현(목표갯수×기준단가, dkDailyTargetAmount 파생).
+ * actualAmount = SF Formula `DailyActualSalesAmount__c`(총 실적 = 대표금액+기타금액) 재현(dailyTotalActualSalesAmount 파생).
+ * professionalPromotionTeam = 조원일정(투입 당시) 값. isWorkReport/commuteDate 는 teamMemberSchedule 조인.
  */
 data class PromotionTargetActualReportRow(
     val promotionName: String?,
@@ -54,7 +55,7 @@ data class PromotionTargetActualReportRow(
     val employeeName: String?,
     val professionalPromotionTeam: String?,
     val scheduleDate: String?,
-    val targetAmount: Long?,
+    val targetAmount: BigDecimal?,
     val actualAmount: BigDecimal?,
     val standLocation: String?,
     val primarySalesQuantity: BigDecimal?,
